requirements.txt
PyMuPDF>=1.23
Pillow>=10.0
pronouncing>=0.2
pytest>=8.0
